removed unnecessary method

+++ avr_main/src/menu/widget/sub_menu.rs
@@ -46,7 +46,7 @@ impl<'a> SubMenu<'a> {
     }
 
     fn get_current_index(&self, line: LcdLine) -> usize {
-        let lcd_index = line as u8 as usize;
+        let lcd_index = line as usize;
         let line_index = self.first_line_to_render.get_current();
         let index = lcd_index + line_index;
         index
